+++ branches/Cobalt-Task4624/RTCP/Cobalt/GPUProc/src/cuda/Kernels/CoherentStokesKernel.h
@@ -43,10 +43,10 @@ namespace LOFAR
       struct Parameters : Kernel::Parameters
       {
         Parameters(const Parset& ps);
-        size_t nrTABs;
-        size_t nrStokes;
-        size_t timeIntegrationFactor;
-        size_t timeParallelFactor;
+        unsigned nrTABs;
+        unsigned nrStokes;
+        unsigned timeIntegrationFactor;
+        unsigned timeParallelFactor;
       };
 
       enum BufferType
@@ -56,13 +56,14 @@ namespace LOFAR
       };
 
       CoherentStokesKernel(const gpu::Stream &stream,
-                             const gpu::Module &module,
-                             const Buffers &buffers,
-                             const Parameters &param);
+                           const gpu::Module &module,
+                           const Buffers &buffers,
+                           const Parameters &param);
+
     };
 
-    // Specialization of the KernelFactory for
-    // CoherentStokesKernel
+    //# --------  Template specializations for KernelFactory  -------- #//
+
     template<> size_t
     KernelFactory<CoherentStokesKernel>::bufferSize(BufferType bufferType) const;
